+++ src/records/traits/ActiveRecord.php
@@ -1,0 +1,112 @@
+<?php
+
+/**
+ * @copyright  Copyright (c) Flipbox Digital Limited
+ * @license    https://github.com/flipboxfactory/craft-ember/blob/master/LICENSE
+ * @link       https://github.com/flipboxfactory/craft-ember
+ */
+
+namespace flipbox\ember\records\traits;
+
+use yii\base\InvalidParamException;
+use yii\db\ActiveQuery;
+use yii\db\ActiveQueryInterface;
+
+trait ActiveRecord
+{
+    /**
+     * Returns the named attribute value.
+     * If this record is the result of a query and the attribute is not loaded,
+     * `null` will be returned.
+     * @param string $name the attribute name
+     * @return mixed the attribute value. `null` if the attribute is not set or does not exist.
+     * @see hasAttribute()
+     */
+    abstract public function getAttribute($name);
+    
+    /**
+     * Check whether the named relation has been populated with records.
+     * @param string $name the relation name, e.g. `orders` for a relation defined via `getOrders()` method
+     * (case-sensitive).
+     * @return bool whether relation has been populated with records.
+     * @see getRelation()
+     */
+    abstract public function isRelationPopulated($name);
+
+    /**
+     * Returns the relation object with the specified name.
+     * A relation is defined by a getter method which returns an [[ActiveQueryInterface]] object.
+     * It can be declared in either the Active Record class itself or one of its behaviors.
+     * @param string $name the relation name, e.g. `orders` for a relation defined via `getOrders()` method
+     * (case-sensitive).
+     * @param bool $throwException whether to throw exception if the relation does not exist.
+     * @return ActiveQueryInterface|ActiveQuery the relational query object. If the relation does not exist
+     * and `$throwException` is `false`, `null` will be returned.
+     * @throws InvalidParamException if the named relation does not exist.
+     */
+    abstract public function getRelation($name, $throwException = true);
+
+    /**
+     * Declares a `has-one` relation.
+     * The declaration is returned in terms of a relational [[ActiveQuery]] instance
+     * through which the related record can be queried and retrieved back.
+     *
+     * A `has-one` relation means that there is at most one related record matching
+     * the criteria set by this relation, e.g., a customer has one country.
+     *
+     * For example, to declare the `country` relation for `Customer` class, we can write
+     * the following code in the `Customer` class:
+     *
+     * ```php
+     * public function getCountry()
+     * {
+     *     return $this->hasOne(Country::className(), ['id' => 'country_id']);
+     * }
+     * ```
+     *
+     * Note that in the above, the 'id' key in the `$link` parameter refers to an attribute name
+     * in the related class `Country`, while the 'country_id' value refers to an attribute name
+     * in the current AR class.
+     *
+     * Call methods declared in [[ActiveQuery]] to further customize the relation.
+     *
+     * @param string $class the class name of the related record
+     * @param array $link the primary-foreign key constraint. The keys of the array refer to
+     * the attributes of the record associated with the `$class` model, while the values of the
+     * array refer to the corresponding attributes in **this** AR class.
+     * @return ActiveQueryInterface the relational query object.
+     */
+    abstract public function hasOne($class, $link);
+
+    /**
+     * Declares a `has-many` relation.
+     * The declaration is returned in terms of a relational [[ActiveQuery]] instance
+     * through which the related record can be queried and retrieved back.
+     *
+     * A `has-many` relation means that there are multiple related records matching
+     * the criteria set by this relation, e.g., a customer has many orders.
+     *
+     * For example, to declare the `orders` relation for `Customer` class, we can write
+     * the following code in the `Customer` class:
+     *
+     * ```php
+     * public function getOrders()
+     * {
+     *     return $this->hasMany(Order::className(), ['customer_id' => 'id']);
+     * }
+     * ```
+     *
+     * Note that in the above, the 'customer_id' key in the `$link` parameter refers to
+     * an attribute name in the related class `Order`, while the 'id' value refers to
+     * an attribute name in the current AR class.
+     *
+     * Call methods declared in [[ActiveQuery]] to further customize the relation.
+     *
+     * @param string $class the class name of the related record
+     * @param array $link the primary-foreign key constraint. The keys of the array refer to
+     * the attributes of the record associated with the `$class` model, while the values of the
+     * array refer to the corresponding attributes in **this** AR class.
+     * @return ActiveQueryInterface the relational query object.
+     */
+    abstract public function hasMany($class, $link);
+}