+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
 - ArrayHelper to assist with filtering null/empty array values
 - CacheableActiveQuery, similar to an ElementQuery
 - FieldLayoutHelper to assist with resolving a FieldLayout
+- Site and FieldLayout traits
 
 ### Changed
 - Exceptions now inherit Yii's base Exception, not ErrorException